new polygon drawing code
instead of using GL_INVERT, we increment/decrement the pixels so that we can
detect nonzero areas. this means we can now handle overlapping polygons.
optionally, you can switch back to even-odd filling if desired (needs to be
exposed to the stylesheet).

in the same turn, it switches the clipping from the depth to the stencil buffer
so that we don't need to request a depth buffer at all. prior to rendering,
it masks the 0x80 bit to indicate the current tile extent.

it also switches to front-to-back compositing, meaning that things drawn later
in a frame will appear *behind* what is already visible. this allows us to use
the stencil buffer to mask areas that are fully opaque so that we can cull
fragments in those areas early on.

fixes #55
fixes #77
fixes #177