compare bboxes before calculating rotation range
The bboxes of some of the glyphs returned by the rtree may
not intersect with the bbox of the glyph it is calculating
the range for. By calculating the bboxes for the minimum
scale they might intersect, and comparing those, it is possible
to skip rotation range calculations for a significant number
of glyphs.

The performance impact depends on the number of horizontal labels
in the map.

This decreased placement time by 15% for the default debug style
and by 70% for a style with more horizontal labels.