StyleLayer
Introduce a StyleLayer class for the in-memory representation
of a style layer.

It handles constant resolution, paint property cascading, and
paint property calculation. The paint functions now accept a
layer as one of the parameters and read all layer properties from
it. This simplifies the rendering flow quite a bit.

The concept of a "bucket" is now something that only the workers
need to know about. The main thread does not create buckets, it
merely stores the buffers and elementGroups on the tile when they
are received from the worker.