fix dashed lines in overscaled tiles
For overscaled tiles we were also scaling units used to measure the
distance from point to point. For really overscaled tiles with long
segments the length of a single segment sometimes exceeded the maximum
distance we have space for in the buffers.

The fix is to scale down the dash size instead of scaling up the
distance.

fixes #1132