switch geojson units from 4096 to pixels (breaking)
The units for these GeoJSONSource options changed from 4096 to pixels:
- tolerance
- buffer
- clusterSize

To convert an old value to pixels divide by 8.
An old value of 400 is the same as 400 / 8 = 50 now.

This also increases the default buffer size (how much data padding
exists on the sides of the tiles) to 128 pixels. Having a big enough
buffer is important for crosstile label placement and to avoid clipping
for at tile edges for symbols that are allowed to overlap with
each other. This does make the tiles bigger but it think it's worth it.