remove `flip` from LineBucket.addLine
`flip` used to be used to changed the directions of extrudes after a
`flipbevel` vertex. Vertices are made up of two extruded points: one for
the left side, one for the right side. After a `flipbevel` vertex the
order of these extrudes would need to flip (right side before the left
side).

f5b58ef25b9e54d90f5eb1028b118b1c7ad71831 changed this by just adding an
extra vertex. This was a great idea. While it does add extra vertices,
`flipbevel` vertices area rare so the number is negligible.

After that commit `flip` was always 1. This just removes it.

I'm glad to see this go. It's caused so many headaches.