Fix an issue where 'in' '$type' drops features
If 'Point' is listed inside an 'in' facet on the '$type' property as the
last value when all geometry types are listed, it will be ignored.

`indexOf` accepts multiples params and `map` emits multiple params. Only
the first param from map should be passed to indexOf. The order of 'Point'
in the arrays ends up causing the value not to be found and the index in
the compiled filter to be -1.

Using a proper inline function solves the issue.