Annotate reference and documentation with SDK requirements
Each layer type and property is annotated with the first version of the JS, iOS, and Android SDK that supports the feature. For properties, support is divided into basic support and data-driven styling support. Absence of an annotation for a particular SDK implies that the feature is unsupported. Unsupported combinations are marked as such and in some cases linked to the relevant GitHub issue.

Most features are annotated as supported starting with GL JS 0.10.0, iOS SDK 2.0.0, and Android SDK 2.0.1 -- the first stable releases that supported GL rendering of the initial v8 style specification. The exceptions are the following:

Source support:
* Image and Video sources
* GeoJSON clustering

Properties added to v8 after the initial release:
* line-offset
* text-pitch-alignment
* icon-text-fit
* icon-text-fit-padding

Properties where data-driven styling support has been introduced in GL JS:
* circle-opacity
* circle-blur
* fill-color
* fill-outline-color
* circle-color
* circle-radius

Filter operators are not yet annotated. The interesting operator is `has`/`!has`, which was a recent introduction.