preserveDrawingBuffer documentation addition (#2813)
Added note that Firefox will not print the map unless this is set to true (longstanding/unassigned [Firefox bug](https://bugzilla.mozilla.org/show_bug.cgi?id=976173)). Not sure if this is the appropriate place for a browser specific note or not, but it might save someone some hair pulling.