Allow drag rotate handler to control pitch (#3105)
* Allow drag rotate handler to control pitch

Right-click and drag previously only controlled the bearing of the map.
Now it also controls the pitch along with bearing.

Dragging up and down controls the pitch. Dragging left and right controls
the bearing.

The pitch control is opt-out and can be disabled by passing `pitchWithRotate:
false` when initializing the Map. The experience can be demoed at
http://localhost:9966/debug/

* Remove '.mason' from ignore files