Stabilize the map-text-depthtest render test (#4219)
The feature in this test is at an angle of 180 degrees (straight
vertical).  Because we use various multiples of 90 degrees in
SymbolBucket to drop incorrectly oriented glyphs
(https://github.com/mapbox/mapbox-gl-js/blob/master/src/data/bucket/symbol_bucket.js#L536-L542),
the orientation of the rendered text is unstable with respect to changes
in rounding based on how each symbol's `glyphAngle` is encoded.