Fix off-by-one error in sprite atlas wrapping
For patterned images, the sprite atlas buffers a row of pixels so that interpolation is continuous at the edges; the bottom row is copied to the top, right to left, and so on. But the clever indexing was too clever, and copied the top row from the next row down, rather than from the bottom.