fix the x component of text-offset for labels along lines
This also changes how we iterate along the line to find the correct
point. The previous algorithm sorted the quads by offset (during bucket
creation) and then iterated along the line a single time, continuing
from the previous point.

The new approach is simpler. It iterates for each glyph independently
and doesn't require them to be sorted. It's a tiny bit slower but it's
also easier to work with. For example, implementing text-offset changes
the sorted order you'd need for the previous algorithm.

As we explore other improvements like #4825 #4826 this will be easier to
work with. We can always optimize again later when things settle.