Enable property functions for line-width (#4773)
line-dasharray is scaled by line-width and therefore requires the value of line-width at floor(current z). Therefore in order to enable property functions for line-width, we add an extra phantom paint attribute (which copies the line-width expression, but evaluates it at floor(z)) so that we can move the scaling calculations to the shader.