Fix packaging issues with style-spec package
Fixes #5502

- The prepublish step now:
  - Compiles the style-spec module to `dist/index.js`. Note that this is still a *Node* module: it does not include external dependencies or browser implementations of built-in node modules.
  - Copies `flow-typed/` from the root mapbox-gl-js repo. As @davidtheclark mentions in https://github.com/mapbox/mapbox-gl-js/issues/5502#issuecomment-338691744, this may not be sufficient to make flow types work out of the box, but it at least ensures that the necessary type definitions are distributed with the style-spec package.

The majority of users should simply be able to do
`require("@mapbox/mapbox-gl-style-spec")`, and it should work fine both in
Node and standard browser bundles that understand commonjs.  But it's also possible to require individual modules like `require("@mapbox/mapbox-gl-style-spec/expression")`, so long as the client either (a) uses `flow-remove-types` and Node >= 6.0, or (b) respects the transforms `browserify` field of package.json.