Fix incorrect textPixelRatio argument (#6135)
* Add failing test for issue #6075

* Fix incorrect textPixelRatio argument

Closes #6075

`textPixelRatio` is used to scale tile coordinates to viewport
coordinates; the value that was being passed here, EXTENT / tileSize,
was the reciprocal of the ratio needed -- we were passing 16 instead of
0.0625.

As a result, the corners of the viewport-projected collision box that we
reconstructed here,
https://github.com/mapbox/mapbox-gl-js/blob/498215e8b2e289dd21f1c471249d16f2ad675151/src/symbol/collision_index.js#L322-L325,
were 16^2 times farther from the anchor than they should have been. With
`icon-offset: [0, 0]`, this didn't matter, because the incorrectly large
box would contain the (much smaller) correct box.  But when there's an
offset, it too gets scaled too much, and so we run out of luck.