Fix flickering 3d terrain #31
The cause was in broken logic of _findTileCoveringTileID: in one of the paths it didn't check if returned tile is loaded (as it should lookup for renderable tiles only).
Moved proxy source update from updateTileBinding (happens on every render frame) to update (called when all other sources are updated).

Couldn't figure out economical way to make useful and reproducile test for this. Instead assert inconsistencies in lookup cache. Measured and this even seems beneficial (as proxy cache is not reset on every render call).

Fixes: https://github.com/mapbox/mapbox-gl-js-internal/issues/31