optimize speedindex recording (#274)
readPixels can take a fairly long time to run. Instead of running it on
every frame that we save, copy the frame to a texture and read it later.
This reduces the effect of speedindex on other metrics.

Previously ~200ms of cpu time was spent reading pixels. Now it's close
to 0 in chrome and ~20ms in Firefox. In Chrome, gpu timing shows no
significant gpu time spent copying it to a texture.