Fix terrain flickering using off-by-one stencil value (internal-330)
* Fix terrain flickering using off-by-one stencil value

The gl spec mention the following for glStencilFunc:

>ref
>Specifies the reference value for the stencil test.
>ref is clamped to the range [0,2n−1], where n is the
number of bitplanes in the stencil buffer. The initial value is 0.

We're using a combined depth/stencil buffer with 8 bits
dedicated to stencil, which means values will wrap to 0
when being set to 256 leading to an incorrect stencil test,
stencil values must remain within the range [0..255].

Fixes internal-288

* Revert unnecessary change