Terrain: Don't load unnecessary tiles (#10444)
* Don't load unecessary tiles

Initial cover for sources other than DEM, before all the  DEM tiles get loaded, shouldn't consider all the tiles with aabb.z range of [-10000, 10000]. As viewing frustum stretches all the way to the sea level, this caused that, when on terrain many unecessary tiles would get requested, even not contributing to the scene.

Instead, consider aabb.z in range { min:centerElevation, max:centerElevation}. Once DEM tiles get loaded, the tile gets proper bounds.

* Simplify sanfrancisco-terrain-draped tests

The reason the test was working the way it was (wasn't working on gl-native though) was in the same issue this PR fixes: tiles of zoom level 15 were fetched and retained because of "deep" z range. When zoom was reduced to zoom 15 they were available.
Dropped the "load at 15.2 and zoom out to 15.0" approach to retain the required tiles. Moved the center to a position where tiles 15 would get loaded on terrain with zoom 15. The test now corresponds to behavior on real map.

Updated expectations, -movelayer variant shows the contour under as expected.