use different attribute names for dashes to avoid collision
The attribute names used for dashes were the same as those used for
patterns. This was leading to incorrectly generated shaders.

Since there was no pattern set for the dash layer it was setting the
defines that generate the shader without those attributes.

The line atlas texture is now bound with gl.REPEAT to render more than
one dash.