[MAPS3D-1467] Fix precision related flickering on far plane (internal-1876)
The high resolution of depth buffer is not required for very high zooms (e.g. 19) while the flickering effects are possible on sources with low max zoom. Anyway, when tilting the map to show horizon, a way further far distance is used with no problems so there should be no issues for doing the same in this case.